backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
sentence-transformers==2.5.1
PyPDF2==3.0.1
requests==2.31.0
python-docx==1.1.0
aiohttp==3.9.3
beautifulsoup4==4.12.3
python-multipart==0.0.9
langdetect==1.0.9
brotli==1.1.0
slowapi==0.1.9
bleach==6.1.0
python-magic-bin==0.4.14; sys_platform == 'win32'
python-magic==0.4.27; sys_platform != 'win32'
